requirements.txt
Flask
scikit-learn
numpy
Pillow
gunicorn # Optional, but good for running Flask in production